chore: handle default in structurer for TiDB
